move hook that still used out of Fractal folder
